requirements.txt
tweepy>=4.14.0
instagrapi>=2.1.0
pandas>=2.0.0
schedule>=1.2.0
google-generativeai>=0.8.0
requests>=2.31.0
